Ability to use term slug instead of id in conditional taxonomy
In cases where custom fields are exported to PHP code and custom
taxonomy terms are also in code it's necessary to be able to use the
slug instead of ID as the ID in your local copy of site may be
different then on production server. First attempted to use get_term_by
in the register field group function to convert slug to ID but the
order of hooks involved prohibits this as that function doesn't return
anything.