Added support to textarea rows number.
Size matters, this is a Design principle. When you display a large space, user can think that they need to fill all that space. Sometimes this cause confusion, when you have a postal code field, and this field is long, if you don't know what is a postal code, you will think it's a long information, not short.

So the size os the fields in all ACF needed to be adjustable, but to simplify, this is the most usefull I think. We can select 2 rows and tell the client, write 2 lines of information, is enough. He will see, the space of the textarea is for 2 lines.

I hope I could be easy to understand. Sorry about my english!